Serious refactoring of tests
Including, splitting into manageable chunks across multiple files.

Dividing tests into more test classes, doing this revealed a number of tests that were passing previously be virtue of the state that the test class happned to be in. These issues have now been corrected.

Lots of whitespace, pep8, line-length fixes.

Imports rationalized.
